Adding SSL support
Enabling SSL support, requiring it for the `aptible` user. Forcing/using SSL
with MySQL is a little wonky; in 5.6 it's essentially a request from both the
client and server side so long as the server decides it can support it. Because
of this, I'm also adding a `-nossl` user that doesn't require SSL as a
workaround for cases where clients might not fully support SSL, since otherwise
an old/broken client might be unable to connect.